Update css petLists
