+++ src/Pet_Page/PetUpdate.js
@@ -1,174 +1,121 @@
-import { useState } from "react";
-import { useNavigate } from "react-router-dom";
-import axios, { BASE_URL } from '../services/axios';
-import { useEffect } from "react";
-import '../styles/petslist.scss';
-import { useCallback } from "react";
-const PetsList = () => {
-    const [pets, setPets] = useState([]);
-    const [currentPage, setCurrentPage] = useState(1);
-    const [petsPerPage] = useState(8);
-    const [searchParams, setSearchParams] = useState({
-        name: "",
-        age: "",
-        categoryID: 0,
-        sex: ""
-    });
-    const navigate = useNavigate();
-    const [ageError, setAgeError] = useState("");
-    const [noResults, setNoResults] = useState(false);
+import React, { useState } from 'react';
+import ShibaImage from '../assets/images/Shiba.jpg';  // Path to your image
+import "../styles/pet.scss";
+const Pets = () => {
+  const [name, setName] = useState('');
+  const [breed, setBreed] = useState('');
+  const [age, setAge] = useState('');
+  const [weight, setWeight] = useState('');
+  const [gender, setGender] = useState('');
+  const [spayed, setSpayed] = useState(false);
+  const [vaccinated, setVaccinated] = useState(false);
+  const [rabiesvaccinated, setRabiesvaccinated] = useState(false);
+  const [toiletTrained, setToiletTrained] = useState(false);
+  const [friendly, setFriendly] = useState(false);
+  const [diet, setDiet] = useState('');
 
-    useEffect(() => {
-        const checkRole = async () => {
-            const roleID = localStorage.getItem("roleID");
-            try {
-                if (roleID === "3") {
-                    await apiListPets(); 
-                }
-            } catch (error) {
-                console.error("Error pets:", error);
-            }
-        };
-        checkRole();    
-    }, [navigate]);
+  const handleUpdate = () => {
+    const updatedData = {
+      name, breed, age, weight, gender, spayed, vaccinated, toiletTrained, rabiesvaccinated, friendly, diet,
+    };
+    console.log("Updated User Info:", updatedData);
+  };
 
-// Gọi API lấy danh sách các pet cho người dùng
-const apiListPets = async () => {
-    try {
-        const response = await axios.get("/pets/showListOfPets");
-        setPets(response.data);
-    } catch (error) {
-        console.error("Error Api pets:", error);
-        if (error.code === 'ERR_NETWORK') {
-            console.error("Network error!");
-        }
-     
-    }
-};
-// Hàm xử lý khi người dùng nhập thông tin tìm kiếm
-const handleSearch = async (e) => {
-    e.preventDefault();
-    try {
-        const searchData = {
-            name: searchParams.name || "",
-            age: searchParams.age ? parseFloat(searchParams.age) : 0,
-            categoryID: searchParams.categoryID || 0,
-            sex: searchParams.sex || ""
-        };
-
-        const response = await axios.get('/pets/searchByNameAndBreed', { params: searchData });
-        if (response.data.length === 0) {
-            setNoResults(true);
-            setPets([]);
-            } else {
-            setNoResults(false);
-            setPets(response.data);
-            }
-    } catch (error) {
-        console.error("Error searching pets:", error);
-        setNoResults(true);
-        setPets([]);
-      }
-};
-
-const handleInputChange = (e) => {
-    const { name, value } = e.target;
-    if (name === 'age') {
-        const floatValue = parseFloat(value);
-        if (value === "" || (floatValue >= 1 && !isNaN(floatValue))) {
-            setSearchParams(prevParams => ({
-                ...prevParams,
-                [name]: value
-            }));
-            setAgeError("");
-        } else {
-            setAgeError("Age must be 1 or greater");
-        }
-    } else {
-        setSearchParams(prevParams => ({
-            ...prevParams,
-            [name]: name === 'categoryID' ? (value === '' ? 0 : parseInt(value)) : value
-        }));
-    }
-};
-// Hàm xử lý khi người dùng click vào một pet
-    const handlePetClick = useCallback((petID) => {
-        navigate(`/pet/${petID}`);
-    }, [navigate]);
-// Hàm lấy đường dẫn ảnh từ API
-    const getImageUrl = useCallback((imgUrl) => {
-        if (!imgUrl) return '/path/to/default/image.jpg';
-        if (imgUrl.startsWith('images\\')) return `${BASE_URL}${imgUrl.replace('\\', '/')}`;
-        return imgUrl;
-    }, []);
-
-    const indexOfLastPet = currentPage * petsPerPage;
-    const indexOfFirstPet = indexOfLastPet - petsPerPage;
-    const currentPets = pets.slice(indexOfFirstPet, indexOfLastPet);
-
-    const paginate = (pageNumber) => setCurrentPage(pageNumber);
-    return (
-        <div className="pets-list-container">
-            
-            <form onSubmit={handleSearch} className="search-form">
-                <input
-                    type="text"
-                    name="name"
-                    value={searchParams.name}
-                    onChange={handleInputChange}
-                    placeholder="Name"
-                />
-                <input
-                    type="number"
-                    name="age"
-                    value={searchParams.age}
-                    onChange={handleInputChange}
-                    placeholder="Select Age"
-                    step="1"
-                    min="1"
-                />
-                {ageError && <span className="error">{ageError}</span>}
-                
-                <select 
-                    name="categoryID"
-                    value={searchParams.categoryID}
-                    onChange={handleInputChange}
-                >
-                    <option value={0}>Select Category</option>
-                    <option value={1}>Dog</option>
-                    <option value={2}>Cat</option>
-                </select>
-                <select 
-                    name="sex"
-                    value={searchParams.sex}
-                    onChange={handleInputChange}
-                >
-                    <option value="">All</option>
-                    <option value="male">Male</option>
-                    <option value="female">Female</option>
-                </select>
-                <button type="submit" disabled={ageError !== ""}>Search</button>
-            </form>
-            <div className="pets-grid">
-                {currentPets.length > 0 ? currentPets.map((pet) => (
-                    <div key={pet.petID} className="pet-item" onClick={() => handlePetClick(pet.petID)}>
-                        <img src={getImageUrl(pet.img_url)} alt={pet.name} />
-                        <h3>{pet.name}</h3>
-                        <p>Age: {pet.age}</p>
-                        <p>Sex: {pet.sex}</p>
-                        <button onClick={() => handlePetClick(pet.petID)}>View Details</button>
-                    </div>
-                )) : <p>No pets found</p>}
-            </div>
-            <div className="pagination">
-                {Array.from({ length: Math.ceil(pets.length / petsPerPage) }, (_, i) => (
-                    <button key={i} onClick={() => paginate(i + 1)}>
-                        {i + 1}
-                    </button>
-                ))}
-            </div>
+  return (
+    <div className="container pets-container">
+      <div className="row">
+        {/* Left Side with Image */}
+        <div className="col-lg-6 col-md-6 col-sm-12 text-center">
+          <img src={ShibaImage} alt="Shiba" className="img-fluid mb-3 photo" />
+          <button className="pet-btn w-100 mb-3 upload">Upload</button>
         </div>
-    )
-}
 
-export default PetsList;
+        {/* Right Side with Form */}
+        <div className="col-lg-6 col-md-6 col-sm-12 text-center">
+          {/* Name Input */}
+          <div className="row mb-3">
+            <div className="col-sm-8 mx-auto">
+              <input type="text" placeholder="Name" className="form-control" value={name} onChange={(e) => setName(e.target.value)} />
+            </div>
+          </div>
+
+          {/* Breed Input */}
+          <div className="row mb-3">
+            <div className="col-sm-8 mx-auto">
+              <input type="text" placeholder="Breed" className="form-control" value={breed} onChange={(e) => setBreed(e.target.value)} />
+            </div>
+          </div>
+
+          {/* Age Input */}
+          <div className="row mb-3">
+            <div className="col-sm-8 mx-auto">
+              <input type="number" placeholder="Age" className="form-control" value={age} onChange={(e) => setAge(e.target.value)} />
+            </div>
+          </div>
+
+          {/* Weight Input */}
+          <div className="row mb-3">
+            <div className="col-sm-8 mx-auto">
+              <input type="number" placeholder="Weight" className="form-control" value={weight} onChange={(e) => setWeight(e.target.value)} />
+            </div>
+          </div>
+
+          {/* Gender Select */}
+          <div className="row mb-3">
+            <div className="col-sm-8 mx-auto">
+              <select className="form-select" value={gender} onChange={(e) => setGender(e.target.value)}>
+                <option value="" disabled hidden>Choose Gender</option>
+                <option value="Male">Male</option>
+                <option value="Female">Female</option>
+              </select>
+            </div>
+          </div>
+
+          {/* Diet Input */}
+          <div className="row mb-3">
+            <div className="col-sm-8 mx-auto">
+              <input type="text" placeholder="Diet" className="form-control" value={diet} onChange={(e) => setDiet(e.target.value)} />
+            </div>
+          </div>
+
+          {/* Checkboxes for Attributes */}
+          <div className="row">
+            <div className="col-md-6">
+              <div className="form-check mb-3">
+                <input className="form-check-input" type="checkbox" checked={spayed} onChange={(e) => setSpayed(e.target.checked)} />
+                <label className="form-check-label">Spayed</label>
+              </div>
+
+              <div className="form-check mb-3">
+                <input className="form-check-input" type="checkbox" checked={toiletTrained} onChange={(e) => setToiletTrained(e.target.checked)} />
+                <label className="form-check-label">Toilet Trained</label>
+              </div>
+
+              <div className="form-check mb-3">
+                <input className="form-check-input" type="checkbox" checked={rabiesvaccinated} onChange={(e) => setRabiesvaccinated(e.target.checked)} />
+                <label className="form-check-label">Rabies Vaccinated</label>
+              </div>
+            </div>
+
+            <div className="col-md-6">
+              <div className="form-check mb-3">
+                <input className="form-check-input" type="checkbox" checked={vaccinated} onChange={(e) => setVaccinated(e.target.checked)} />
+                <label className="form-check-label">Vaccinated</label>
+              </div>
+
+              <div className="form-check mb-3">
+                <input className="form-check-input" type="checkbox" checked={friendly} onChange={(e) => setFriendly(e.target.checked)} />
+                <label className="form-check-label">Friendly</label>
+              </div>
+            </div>
+          </div>
+
+          {/* Update Button */}
+          <button className="pet-btn w-50" onClick={handleUpdate}>Update</button>
+        </div>
+      </div>
+    </div>
+  );
+};
+
+export default Pets;
