+++ src/styles/petslist.scss
@@ -3,15 +3,9 @@
   max-width: 1200px;
   margin: 0 auto;
 
-  h1 {
-    text-align: center;
-    margin-bottom: 30px;
-    font-size: 2em;
-    color: #D9293E; /* Thêm màu sắc để tiêu đề nổi bật */
-  }
-
   .search-form {
     display: flex;
+    flex-direction: column;
     justify-content: center;
     align-items: center;
     flex-wrap: wrap;
@@ -57,14 +51,26 @@
     flex-direction: column;
     align-items: center;
     padding: 15px;
-    border: 1px solid #ddd;
+    border: 1px solid #e0e0e0;
     border-radius: 10px;
-    transition: box-shadow 0.3s, transform 0.3s;
+    transition: all 0.3s ease;
     cursor: pointer;
-    background-color: #f6f6f6;
+    background-color: #F8F8F8;
+    background-image: radial-gradient(#f1cf7e 10%, transparent 11%),
+                      radial-gradient(#f1cf7e 10%, transparent 11%);
+    background-size: 10px 10px;
+    background-position: 0 0, 5px 5px;
+    background-repeat: repeat;
+    box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
+    
     &:hover {
-      box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
-      transform: scale(1.03);
+      transform: translateY(-5px);
+      box-shadow: 0 10px 20px rgba(0, 0, 0, 0.15);
+    }
+
+    img, h3, p, button, .pet-info-divider {
+      position: relative;
+      z-index: 2;
     }
 
     img {
@@ -73,37 +79,60 @@
       object-fit: cover;
       border-radius: 10px;
       margin-bottom: 15px;
+      transition: transform 0.3s ease;
+
+      &:hover {
+        transform: scale(1.05);
+      }
     }
 
     h3 {
       margin-bottom: 10px;
-      font-size: 1.2em;
+      font-size: 1.3em;
       font-weight: bold;
       color: #333;
+      text-align: center;
+    }
+
+    .pet-info-divider {
+      width: 40px;
+      height: 3px;
+      background-color: #f1cf7e;
+      margin: 10px auto;
+      transition: width 0.3s ease;
+    }
+
+    &:hover .pet-info-divider {
+      width: 60px;
     }
 
     p {
-      font-size: 0.9em;
-      color: #000000;
+      font-size: 0.95em;
+      color: #666;
       margin: 5px 0;
+      text-align: center;
     }
 
     button {
-      margin-top: 10px;
-      padding: 8px 16px;
+      margin-top: 15px;
+      padding: 10px 20px;
       background-color: #f1cf7e;
-      color: white;
+      color: #333;
       border: none;
       border-radius: 25px;
       cursor: pointer;
-      transition: background-color 0.3s;
+      transition: all 0.3s ease;
+      font-weight: bold;
+      box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
 
       &:hover {
-        background-color: #f1cf7e;
+        background-color: #e6be6a;
+        transform: translateY(-2px);
+        box-shadow: 0 4px 8px rgba(0, 0, 0, 0.15);
       }
     }
   }
-
+  
   .pagination {
     display: flex;
     justify-content: center;
@@ -124,9 +153,113 @@
       }
 
       &.active {
-        background-color: #D9293E;
+        
         color: white;
-        border-color: #D9293E;
+        
+      }
+    }
+  }
+}
+.filter-and-search {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  margin-bottom: 30px;
+  width: 100%;
+
+  .search-section {
+    display: flex; 
+    align-items: center;
+    justify-content: space-between;
+    width: 65%; 
+    margin-bottom: 15px;
+
+    .search-inputs {
+      width: 79%; 
+      input {
+        width: 100%;
+        padding: 12px;
+        border: 1px solid #ccc;
+        border-radius: 10px; 
+        font-size: 1em;
+        box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1); 
+        transition: box-shadow 0.3s ease, border-color 0.3s ease;
+        
+        &:focus {
+          border-color: #e6be6a;
+          box-shadow: 0 0 12px rgba(199, 174, 177, 0.3);
+          outline: none;
+        }
+      }
+    }
+
+    .search-btn {
+      padding: 10px 15px;
+      background-color: #f1cf7e;
+      color: white;
+      border: none;
+      border-radius: 25px;
+      cursor: pointer;
+      box-shadow: 0 4px 12px rgba(0, 0, 0, 0.1); 
+      transition: background-color 0.3s, box-shadow 0.3s, transform 0.3s;
+      width: 18%;
+      &:hover {
+        background-color: #e6be6a; 
+        box-shadow: 0 6px 16px rgba(0, 0, 0, 0.2); 
+        transform: translateY(-2px); 
+      }
+
+      &:active {
+        transform: translateY(0); 
+      }
+    }
+  }
+
+  .search-form {
+    width: 65%;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    margin-left: 19px;
+    .other-filters {
+      display: flex;
+      justify-content: space-between;
+      align-items: center;
+      width: 100%;
+      flex-wrap: wrap;
+
+      input, select, button {
+        flex: 1;
+        margin: 10px;
+        padding: 12px;
+        border: 1px solid #ccc;
+        border-radius: 10px; 
+        transition: box-shadow 0.3s ease, border-color 0.3s ease;
+
+        &:hover {
+          border-color: #e6be6a; 
+          box-shadow: 0 4px 12px rgba(0, 0, 0, 0.1); 
+        }
+      }
+    }
+
+    .filter-section {
+      display: flex;
+      align-items: center;
+      justify-content: center;
+      margin-right: 10px;
+      
+      h3 {
+        display: flex;
+        align-items: center;
+        font-size: 1.2em;
+        color: #413838;
+        font-weight: bold;
+        margin: 0;
+
+        svg {
+          margin-right: 5px;
+        }
       }
     }
   }
